Added: RecipientStage
This allows for variables to be processed for each recipient. JSON cannot be used inside of a recipient section, but a recipient section may be surrounded by JSON.

This also implemented the %n for newline change. All formats and sections are read, have newlines replaced with spaces, then %n replaced with a newline, allowing finer-grain control for template writers.