Updated: Change how getAllChannelNames/Tags work
I moved the old getAllChannelNames/Tags to getAllDefinedChannelNames/Tags. Those methods return a List of all names/tags defined in the loaded config.yml.

The new getAllChannelNames/Tags methods now return all names/tags that are defined or loaded in a Set, to prevent duplicates. This also changes how the getChannelByTag method works. It first polls the defined channels, and if a channel could not be found, it then polls the loaded channels.